@@ -21,108 +22,111 @@
   },
   "dependencies": {
     "classnames": "2.2.5",
-    "clipboard": "^1.6.1",
-    "codemirror": "5.26.0",
-    "color": "1.0.3",
-    "core-js": "2.4.1",
+    "clipboard": "1.7.1",
+    "codemirror": "5.29.0",
+    "color": "2.0.0",
+    "core-js": "2.5.1",
     "crypto-js": "3.1.9-1",
     "file-saver": "1.3.3",
-    "golden-layout": "1.5.8",
-    "inversify": "4.1.1",
+    "golden-layout": "1.5.9",
+    "inversify": "4.3.0",
     "inversify-inject-decorators": "3.0.1",
     "isomorphic-fetch": "2.2.1",
     "jed": "1.1.1",
     "jquery": "3.2.1",
     "keymaster": "1.6.2",
     "lodash": "4.17.4",
-    "markdown-it": "8.3.1",
+    "markdown-it": "8.4.0",
     "moment": "2.18.1",
     "po2json": "0.4.5",
-    "react": "15.6.0",
-    "react-bootstrap": "0.31.0",
-    "react-dom": "15.6.0",
-    "react-router": "4.1.1",
-    "react-router-bootstrap": "0.24.2",
-    "react-router-dom": "4.1.1",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
+    "react-router": "4.2.0",
+    "react-router-dom": "4.2.2",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.0",
+    "rxjs": "5.4.3",
     "spark-md5": "3.0.0",
     "tslib": "1.7.1",
     "tv4": "1.3.0"
   },
   "devDependencies": {
-    "@types/chai": "4.0.0",
-    "@types/chai-as-promised": "0.0.31",
-    "@types/classnames": "2.2.0",
-    "@types/clipboard": "^1.5.33",
-    "@types/codemirror": "0.0.40",
-    "@types/color": "1.0.4",
-    "@types/core-js": "0.9.41",
-    "@types/crypto-js": "3.1.33",
-    "@types/enzyme": "2.8.0",
-    "@types/esprima": "2.1.33",
+    "@types/chai": "4.0.4",
+    "@types/chai-as-promised": "7.1.0",
+    "@types/classnames": "2.2.3",
+    "@types/clipboard": "1.5.34",
+    "@types/codemirror": "0.0.46",
+    "@types/color": "2.0.0",
+    "@types/core-js": "0.9.43",
+    "@types/crypto-js": "3.1.35",
+    "@types/enzyme": "2.8.7",
+    "@types/esprima": "4.0.1",
     "@types/file-saver": "0.0.1",
-    "@types/glob": "5.0.30",
-    "@types/history": "4.5.2",
+    "@types/glob": "5.0.32",
+    "@types/history": "4.6.0",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/jquery": "2.0.46",
+    "@types/jquery": "3.2.12",
     "@types/keymaster": "1.6.28",
-    "@types/lodash": "4.14.66",
-    "@types/markdown-it": "0.0.2",
-    "@types/mocha": "2.2.41",
-    "@types/node": "7.0.31",
-    "@types/react": "15.0.28",
-    "@types/react-addons-test-utils": "0.14.18",
-    "@types/react-bootstrap": "0.0.50",
-    "@types/react-dom": "15.5.0",
-    "@types/react-router": "4.0.11",
-    "@types/react-router-bootstrap": "0.24.1",
-    "@types/react-router-dom": "4.0.4",
+    "@types/lodash": "4.14.74",
+    "@types/markdown-it": "0.0.4",
+    "@types/mocha": "2.2.43",
+    "@types/node": "8.0.28",
+    "@types/react": "16.0.5",
+    "@types/react-addons-test-utils": "0.14.19",
+    "@types/react-dom": "15.5.4",
+    "@types/react-router": "4.0.15",
+    "@types/react-router-dom": "4.0.7",
     "@types/tv4": "1.2.28",
-    "archiver": "1.3.0",
-    "awesome-typescript-loader": "3.1.3",
-    "chai": "4.0.2",
-    "chai-as-promised": "7.0.0",
+    "archiver": "2.0.3",
+    "husky": "0.14.3",
+    "lint-staged": "4.2.1",
+    "prettier": "1.7.0",
+    "awesome-typescript-loader": "3.2.3",
+    "chai": "4.1.2",
+    "chai-as-promised": "7.1.1",
     "clean-webpack-plugin": "0.1.16",
     "copy-webpack-plugin": "4.0.1",
-    "css-loader": "0.28.4",
-    "enzyme": "2.8.2",
+    "css-loader": "0.28.7",
+    "enzyme": "2.9.1",
     "esprima": "4.0.0",
-    "extract-text-webpack-plugin": "2.1.2",
+    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "0.11.2",
     "git-revision-webpack-plugin": "2.5.1",
     "glob": "7.1.2",
-    "html-webpack-plugin": "2.28.0",
+    "html-webpack-plugin": "2.30.1",
     "istanbul": "0.4.5",
-    "istanbul-instrumenter-loader": "2.0.0",
-    "istanbul-lib-instrument": "1.7.2",
-    "json-loader": "0.5.4",
-    "karma": "1.7.0",
+    "istanbul-instrumenter-loader": "3.0.0",
+    "istanbul-lib-instrument": "1.8.0",
+    "json-loader": "0.5.7",
+    "karma": "1.7.1",
     "karma-bamboo-reporter": "0.1.2",
-    "karma-chrome-launcher": "2.1.1",
+    "karma-chrome-launcher": "2.2.0",
     "karma-coverage": "1.1.1",
     "karma-firefox-launcher": "1.0.1",
     "karma-mocha": "1.3.0",
-    "karma-mocha-reporter": "2.2.3",
-    "karma-phantomjs-launcher": "1.0.4",
+    "karma-mocha-reporter": "2.2.4",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.3",
-    "mocha": "3.4.2",
+    "karma-webpack": "2.0.4",
+    "mocha": "3.5.3",
     "node-sass": "4.5.3",
-    "pofile": "^1.0.8",
+    "pofile": "1.0.9",
     "raw-loader": "0.5.1",
-    "react-addons-test-utils": "15.5.1",
+    "react-addons-test-utils": "15.6.0",
     "remap-istanbul": "0.9.5",
-    "sass-loader": "6.0.5",
+    "sass-loader": "6.0.6",
     "style-loader": "0.18.2",
-    "ts-loader": "2.1.0",
-    "tslint": "5.4.3",
+    "ts-loader": "2.3.7",
+    "tslint": "5.7.0",
     "tslint-loader": "3.5.3",
-    "typescript": "2.3.4",
+    "typescript": "2.5.2",
     "url-loader": "0.5.9",
-    "webpack": "2.6.1",
-    "webpack-bundle-analyzer": "2.8.2",
-    "webpack-dev-server": "2.4.5",
+    "webpack": "3.6.0",
+    "webpack-bundle-analyzer": "2.9.0",
+    "webpack-dev-server": "2.8.2",
     "webpack-notifier": "1.5.0"
+  },
+  "lint-staged": {
+    "linters": {
+      "*.{ts,tsx,js,json,css,scss}": ["prettier --write", "git add"]
+    }
   }
 }
