@@ -1,27 +1,24 @@
-import { Observable, Subject, Subscription } from 'rxjs/Rx';
+import { Observable, Subject } from "rxjs/Rx";
 import { injectable } from "app/di";
 
 interface INavigationRequest {
-    route: string
+  route: string;
 }
 
 export abstract class INavigationService {
-    abstract navigate(request: INavigationRequest): void
-    abstract getObservable(): Observable<INavigationRequest>
+  abstract navigate(request: INavigationRequest): void;
+  abstract getObservable(): Observable<INavigationRequest>;
 }
 
 @injectable()
 export class NavigationService implements INavigationService {
-    private subject = new Subject<INavigationRequest>();
+  private subject = new Subject<INavigationRequest>();
 
-    constructor() {
-    }
+  navigate(request: INavigationRequest): void {
+    this.subject.next(request);
+  }
 
-    navigate(request: INavigationRequest): void {
-        this.subject.next(request);
-    }
-
-    getObservable(): Observable<INavigationRequest> {
-        return this.subject;
-    }
+  getObservable(): Observable<INavigationRequest> {
+    return this.subject;
+  }
 }
