@@ -1,40 +1,33 @@
 ﻿// Include polyfills to work in IE11
-import 'core-js';
+import "core-js";
 
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+//Expose react and react router in order for golden layout work
+(window as any)["React"] = React;
+(window as any)["ReactDOM"] = ReactDOM;
 
-import { ErrorComponent } from 'app/ui/error.component';
-import { Routes } from 'app/routes';
+import { ErrorComponent } from "app/ui/error.component";
+import { Routes } from "app/routes";
 import { DependecyInjectionSetup } from "app/di-setup";
 
-import 'app/ui/_styles/app.scss';
+import "app/ui/_styles/app.scss";
 
 async function runApp() {
-    const appHostDomElement = document.getElementById('app-container') || document.body;
-    try {
-        await DependecyInjectionSetup.setup()
+  const appHostDomElement = document.getElementById("app-container") || document.body;
+  try {
+    await DependecyInjectionSetup.setup();
 
-        // Render the app
-        ReactDOM.render(
-            <Routes />,
-            appHostDomElement
-        )
-    }
-    catch (ex) {
-        ReactDOM.render(
-            <ErrorComponent
-                headerText="Application failed because of error"
-                error={ex}
-            />,
-            appHostDomElement
-        );
-        throw ex;
-    };
+    // Render the app
+    ReactDOM.render(<Routes />, appHostDomElement);
+  } catch (ex) {
+    ReactDOM.render(<ErrorComponent headerText="Application failed because of error" error={ex} />, appHostDomElement);
+    throw ex;
+  }
 }
 
 Object.assign(window, {
-    app: {
-        run: runApp
-    }
+  app: {
+    run: runApp
+  }
 });
