@@ -3,4 +3,4 @@ import getDecorators from "inversify-inject-decorators";
 
 const container = new Container({ defaultScope: "Singleton" });
 const { lazyInject } = getDecorators(container);
-export { container, lazyInject, inject, injectable }
+export { container, lazyInject, inject, injectable };
