@@ -1,8 +1,8 @@
 export abstract class IAppInfo {
-    gitVersion: string
-    buildVersion: string
-    name: string
-    description: string
-    version: string
-    isDevBuild: boolean
-};
+  gitVersion: string;
+  buildVersion: string;
+  name: string;
+  description: string;
+  version: string;
+  isDevBuild: boolean;
+}
