@@ -2,18 +2,11 @@ import { injectable } from "app/di";
 
 @injectable()
 export class TitleService {
-    constructor(private prefix: string) {
-    }
+  constructor(private prefix: string) {}
 
-    setDocumentTitle(title: string) {
-        if (title) {
-            document.title = this.prefix + "-" + title;
-        }
-        else {
-            document.title = this.prefix;
-        }
-    }
+  setDocumentTitle(title: string) {
+    document.title = title ? this.prefix + "-" + title : this.prefix;
+  }
 }
 
-export interface ITitleService extends TitleService {
-}
+export interface ITitleService extends TitleService {}
