@@ -1,11 +1,11 @@
-import { LoginStatus } from "app/services/login/current-user.provider";
+import { LoginStatus } from "app/services/login/current-user.service";
 import { Observable } from "rxjs/Rx";
 
 export interface IAuthService {
-    init(): Promise<void>;
-    getLoginStatus(): LoginStatus;
-    loginStatusObservable: Observable<LoginStatus>;
-    signOut(): Promise<void>;
-    renderLoginUI(): JSX.Element;
-    initLoginUI(): Promise<void>;
+  init(): Promise<void>;
+  getLoginStatus(): LoginStatus;
+  loginStatusObservable: Observable<LoginStatus>;
+  signOut(): Promise<void>;
+  renderLoginUI(): JSX.Element;
+  initLoginUI(): Promise<void>;
 }
