@@ -1,13 +1,8 @@
 export class UserInfo {
-    id: string
-    attributes: UserAttributes
+  id: string;
+  attributes: UserAttributes;
 }
 
 export class UserAttributes {
-    constructor(
-        public name: string,
-        public email: string,
-        public imageUrl: string,
-    ) {
-    }
+  constructor(public name: string, public email: string, public imageUrl: string) {}
 }
