@@ -1,78 +1,44 @@
-import * as React from 'react';
-import * as cn from 'classnames'
-import { Alert } from "react-bootstrap";
+import * as React from "react";
 
-import './alert-message.component.scss';
+import "./alert-message.component.scss";
 
-type messageType = "info" | "danger" | "warning" | "success";
+type messageType = "primary" | "info" | "danger" | "warning" | "success";
 
 interface IComponentProps {
-    type?: messageType
-    title?: JSX.Element | string
-    message?: JSX.Element | string
-    doNotDetectPreformatted?: boolean
-    onDismiss?: () => void
+  type?: messageType;
+  title?: JSX.Element | string;
+  message?: JSX.Element | string;
+  doNotDetectPreformatted?: boolean;
+  onDismiss?: () => void;
 }
 
-export class AlertMessageComponent extends React.Component<IComponentProps, void> {
-    private getIconClass(type: messageType) {
-        switch (type) {
-            case 'info':
-                return 'glyphicon-info-sign';
-            case 'success':
-                return 'glyphicon glyphicon-ok-sign';
-            case 'danger':
-                return 'glyphicon-exclamation-sign';
-            case 'warning':
-                return 'glyphicon-exclamation-sign';
-        }
+export class AlertMessageComponent extends React.Component<IComponentProps, {}> {
+  render(): JSX.Element | null {
+    const type = this.props.type || "danger";
+    let message = this.props.message || "";
+    let title = this.props.title;
+    const preformatted =
+      typeof message == "string" && !this.props.doNotDetectPreformatted
+        ? message.includes("<") || message.includes("{") || message.includes("[")
+        : false;
+
+    if (type === "danger" && !title && message && message.toString().startsWith("Error: ")) {
+      title = "Error";
+      message = message.toString().substr(7);
     }
 
-    render(): JSX.Element | null {
-        const type = this.props.type || "danger";
-        let message = this.props.message || '';
-        let title = this.props.title;
-        const preformatted = typeof (message) == "string" && !this.props.doNotDetectPreformatted
-            ? (message.includes('<') || message.includes('{') || message.includes('['))
-            : false;
-
-        if (type === "danger" && !title && message && message.toString().startsWith("Error: ")) {
-            title = 'Error';
-            message = message.toString().substr(7);
-        }
-
-        if (message || title) {
-            return <Alert className={`ex-customized-alert`} bsStyle={type} onDismiss={this.props.onDismiss}>
-                <div className="media ex-margin-top-zero">
-                    <div className="media-left">
-                        <span className="text-nowrap">
-                            <span className={`ex-icon-container glyphicon ${this.getIconClass(type)}`} aria-hidden="true"></span>
-                            <span>&nbsp;</span>
-                        </span>
-                    </div>
-                    <div className="media-body ex-vertical-align-middle">
-                        {
-                            title &&
-                            <h4 className="media-heading ex-margin-zero">
-                                <span>
-                                    {title}
-                                </span>
-                            </h4>
-                        }
-                        {
-                            title && message && <p />
-                        }
-                        {
-                            message && (
-                                preformatted
-                                    ? <pre>{message}</pre>
-                                    : message
-                            )
-                        }
-                    </div>
-                </div>
-            </Alert>
-        }
-        return null;
+    if (message || title) {
+      return (
+        <div>
+          <article className={"message is-" + type}>
+            <div className="message-body">
+              {title && <p className="subtitle">{title}</p>}
+              {message && (preformatted ? <pre>{message}</pre> : <p>{message}</p>)}
+            </div>
+          </article>
+        </div>
+      );
     }
+    return null;
+  }
 }
