@@ -1,30 +1,26 @@
-import * as React from 'react';
-import * as cn from 'classnames';
+import * as React from "react";
+import * as cn from "classnames";
 
-import './no-data.component.scss';
+import "./no-data.component.scss";
 
 interface IComponentProps {
-    title: JSX.Element | string
-    description: JSX.Element | string
-    iconClass?: string
-    noBorder?: boolean
+  title: JSX.Element | string;
+  description?: JSX.Element | string;
+  iconClass?: string;
 }
 
-export class NoDataComponent extends React.Component<IComponentProps, void> {
-    render(): JSX.Element {
-        const iconClass = this.props.iconClass || "glyphicon-list";
+export class NoDataComponent extends React.Component<IComponentProps, {}> {
+  render(): JSX.Element {
+    const iconClass = this.props.iconClass || "fa-picture-o";
 
-        return (
-            <div className="row">
-                <div className="col-sm-12">
-                    <br />
-                    <div className={cn("ex-no-data-container", { "no-border": this.props.noBorder })}>
-                        <span className={`no-data-icon glyphicon ${iconClass}`} ></span>
-                        <h2 className="no-data-title">{this.props.title}</h2>
-                        <p className="no-data-description">{this.props.description}</p>
-                    </div>
-                </div>
-            </div>
-        );
-    }
+    return (
+      <div className="no-data-component has-text-centered">
+        <span className="icon is-large">
+          <i className={"fa " + iconClass} aria-hidden="true" />
+        </span>
+        {this.props.title && <h2 className="subtitle">{this.props.title}</h2>}
+        {this.props.description && <p>{this.props.description}</p>}
+      </div>
+    );
+  }
 }
