@@ -1,50 +1,56 @@
-import * as React from 'react';
-import { Link, RouteComponentProps } from 'react-router-dom'
+import * as React from "react";
+import { RouteComponentProps } from "react-router-dom";
 
-import { MainMenuComponent } from 'app/ui/main-menu.component'
-import { PageHeaderComponent } from 'app/ui/_generic/page-header.component';
+import { MainMenuComponent } from "app/ui/main-menu.component";
+import { PageHeaderComponent } from "app/ui/_generic/page-header.component";
 
 import { _T } from "app/services/customizations/localization.service";
 import { lazyInject } from "app/di";
 import { TitleService } from "app/services/infrastructure/title.service";
 import { IAppInfo } from "app/services/infrastructure/app-info";
 
-interface IComponentState {
-}
+interface IComponentState {}
 
-interface IComponentProps extends RouteComponentProps<void> {
-}
+interface IComponentProps extends RouteComponentProps<void> {}
 
 export class AboutComponent extends React.Component<IComponentProps, IComponentState> {
-    @lazyInject(TitleService)
-    private titleService: TitleService;
+  @lazyInject(TitleService) private titleService: TitleService;
+  @lazyInject(IAppInfo) private appInfo: IAppInfo;
 
-    @lazyInject(IAppInfo)
-    private appInfo: IAppInfo;
+  constructor(props: IComponentProps) {
+    super(props);
+    this.state = {};
+    this.titleService.setDocumentTitle(_T("About"));
+  }
 
-    constructor(props: IComponentProps) {
-        super(props);
-
-        this.state = {
-        };
-
-        this.titleService.setDocumentTitle(_T("About"));
-    }
-
-    render(): JSX.Element {
-        return (
-            <div className="container">
-                <MainMenuComponent />
-                <PageHeaderComponent title={_T("About")} />
-                <div className="row">
-                    <div className="col-sm-12">
-                        <p>{this.appInfo.description}</p>
-                        <p><strong>{_T("Package name")}:</strong> {this.appInfo.name}</p>
-                        <p><strong>{_T("App version")}:</strong> {this.appInfo.version}</p>
-                        <p><strong>{_T("Code version")}:</strong> {this.appInfo.gitVersion}</p>
-                    </div>
+  render(): JSX.Element {
+    return (
+      <div className="ex-page-container">
+        <MainMenuComponent />
+        <div className="ex-page-content">
+          <div className="container">
+            <br />
+            <PageHeaderComponent title={_T("About")} />
+            <br />
+            <div className="card">
+              <div className="card-content">
+                <div className="content">
+                  <p>{this.appInfo.description}</p>
+                  <p>
+                    <strong>{_T("Package name")}:</strong> {this.appInfo.name}
+                  </p>
+                  <p>
+                    <strong>{_T("App version")}:</strong> {this.appInfo.version}
+                  </p>
+                  <p>
+                    <strong>{_T("Code version")}:</strong> {this.appInfo.gitVersion}
+                  </p>
                 </div>
+              </div>
             </div>
-        );
-    }
+          </div>
+        </div>
+      </div>
+    );
+  }
 }
