@@ -1,12 +1,10 @@
-import * as React from 'react';
-import { Link, RouteComponentProps } from 'react-router-dom';
+import * as React from "react";
+import { RouteComponentProps } from "react-router-dom";
 
-import { setupActionErrorHandler, callAction } from "app/utils/async-helpers";
+import { callActionSafe } from "app/utils/async-helpers";
 
-import { MainMenuComponent } from 'app/ui/main-menu.component'
-import { PageHeaderComponent } from 'app/ui/_generic/page-header.component';
-import { PageLoadingIndicatorComponent } from 'app/ui/_generic/page-loading-indicator.component';
-import { AlertMessageComponent } from 'app/ui/_generic/alert-message.component';
+import { MainMenuComponent } from "app/ui/main-menu.component";
+import { PageLoadingIndicatorComponent } from "app/ui/_generic/page-loading-indicator.component";
 
 import { lazyInject } from "app/di";
 import { _T } from "app/services/customizations/localization.service";
@@ -14,64 +12,62 @@ import { INotificationService } from "app/services/infrastructure/notification.s
 import { TitleService } from "app/services/infrastructure/title.service";
 import { ILocalizedContentLoader } from "app/services/infrastructure/localized-content-loader";
 
-import './documentation.component.scss'
+import "./documentation.component.scss";
 
 interface IComponentState {
-    isLoading: boolean
-    content: string
+  isLoading: boolean;
+  content: string;
 }
 
-interface IComponentProps extends RouteComponentProps<void> {
-}
+interface IComponentProps extends RouteComponentProps<void> {}
 
 export class DocumentationComponent extends React.Component<IComponentProps, IComponentState> {
-    @lazyInject(INotificationService)
-    private notificationService: INotificationService;
+  @lazyInject(INotificationService) private notificationService: INotificationService;
+  @lazyInject(TitleService) private titleService: TitleService;
+  @lazyInject(ILocalizedContentLoader) private contentLoader: ILocalizedContentLoader;
 
-    @lazyInject(TitleService)
-    private titleService: TitleService;
+  private errorHandler = (err: string) => {
+    this.notificationService.push({ message: err, type: "danger" });
+    this.setState({ isLoading: false });
+  };
 
-    @lazyInject(ILocalizedContentLoader)
-    private contentLoader: ILocalizedContentLoader;
+  constructor(props: IComponentProps) {
+    super(props);
 
-    constructor(props: IComponentProps) {
-        super(props);
+    this.state = {
+      isLoading: true,
+      content: ""
+    };
+  }
 
-        this.state = {
-            isLoading: true,
-            content: ''
-        };
+  async componentDidMount() {
+    this.titleService.setDocumentTitle(_T("Documentation"));
+    await this.loadData();
+  }
+
+  private async loadData() {
+    const content = await callActionSafe(this.errorHandler, async () =>
+      this.contentLoader.getFileContent("reference.html")
+    );
+    if (content) {
+      this.setState({ content: content });
     }
+    this.setState({ isLoading: false });
+  }
 
-    componentDidMount() {
-        this.loadData();
-        this.titleService.setDocumentTitle(_T("Documentation"));
-    }
-
-    private async loadData() {
-        const errorHandler = setupActionErrorHandler((error) => {
-            this.notificationService.push({ type: 'danger', message: error });
-            this.setState({ isLoading: false });
-        })
-
-        const content = await callAction(errorHandler, () => this.contentLoader.getFileContent('reference.html'));
-        if (content) {
-            this.setState({ content: content });
-        }
-        this.setState({ isLoading: false });
-    }
-
-    render(): JSX.Element {
-        return (
-            <div className="ex-page-container">
-                <MainMenuComponent />
-                <div className="ex-scroll-outer container-fluid">
-                    <PageLoadingIndicatorComponent isLoading={this.state.isLoading}>
-                        <div className="doc-section" dangerouslySetInnerHTML={{ __html: this.state.content }}></div>
-                        <br />
-                    </PageLoadingIndicatorComponent>
-                </div>
-            </div>
-        );
-    }
+  render(): JSX.Element {
+    return (
+      <div className="ex-page-container">
+        <MainMenuComponent />
+        <div className="ex-page-content">
+          <div className="container">
+            <PageLoadingIndicatorComponent isLoading={this.state.isLoading}>
+              <div className="doc-section" dangerouslySetInnerHTML={{ __html: this.state.content }} />
+              <br />
+            </PageLoadingIndicatorComponent>
+          </div>
+        </div>
+      </div>
+    );
+  }
 }
