@@ -1,25 +1,208 @@
-import * as React from 'react';
-import { Observable } from 'rxjs';
+import * as React from "react";
+import { ISubscription } from "rxjs/Subscription";
+import { Observable } from "rxjs/Observable";
+import * as keymaster from "keymaster";
 
-import { CodeInputLogoComponent, ICodeInputComponentProps } from 'app/ui/_shared/code-input-logo.component';
+import { as } from "app/utils/syntax-helpers";
+import { lazyInject } from "app/di";
 
-import './code-panel.component.scss'
+import { _T } from "app/services/customizations/localization.service";
+import { Routes } from "app/routes";
+import { ProgramModel } from "app/services/program/program.model";
+import { INotificationService } from "app/services/infrastructure/notification.service";
+import { INavigationService } from "app/services/infrastructure/navigation.service";
+import { ProgramExecutionContext } from "app/services/program/program-execution.context";
+import {
+  ProgramManagementService,
+  IProgramToSaveAttributes,
+  ProgramStorageType
+} from "app/services/program/program-management.service";
+
+import { ShareScreenshotModalComponent } from "app/ui/playground/share-screenshot-modal.component";
+import { CodeInputLogoComponent, ICodeInputComponentProps } from "app/ui/_shared/code-input-logo.component";
+import { SaveProgramModalComponent } from "app/ui/playground/save-program-modal.component";
+import { ProgramControlsMenuComponent } from "app/ui/playground/program-controls-menu.component";
+import { AlertMessageComponent } from "app/ui/_generic/alert-message.component";
+
+import "./code-panel.component.scss";
 
 export interface ICodePanelComponentProps {
-    codeInputProps: ICodeInputComponentProps
+  editorTheme: string;
+  executionService: ProgramExecutionContext;
+  program: ProgramModel;
+  saveCurrentEnabled: boolean;
+  navigateAutomaticallyAfterSaveAs: boolean;
+  externalCodeChanges?: Observable<string>;
+  doNotShowLocalChangesIndicator?: boolean;
 }
 
-export class CodePanelComponent extends React.Component<ICodePanelComponentProps, void> {
+interface IComponentState {
+  isSaveModalActive: boolean;
+  hasLocalTempChanges: boolean;
+  screenshotDataToSave: string;
+  code: string;
+}
 
-    constructor(props: ICodePanelComponentProps) {
-        super(props);
+export class CodePanelComponent extends React.Component<ICodePanelComponentProps, IComponentState> {
+  @lazyInject(INotificationService) private notificationService: INotificationService;
+  @lazyInject(INavigationService) private navigationService: INavigationService;
+  @lazyInject(ProgramManagementService) private managementService: ProgramManagementService;
+  private subscriptions: ISubscription[] = [];
+  private saveTempCodeTimer: any = undefined;
+
+  constructor(props: ICodePanelComponentProps) {
+    super(props);
+    this.state = {
+      isSaveModalActive: false,
+      screenshotDataToSave: "",
+      hasLocalTempChanges: this.props.program.hasTempLocalModifications,
+      code: this.props.program.code
+    };
+  }
+
+  componentDidMount() {
+    this.subscriptions.push(
+      this.props.executionService.onIsRunningChanged.subscribe(() => {
+        // Update state to force a component render
+        this.setState({});
+      })
+    );
+    if (this.props.externalCodeChanges) {
+      this.subscriptions.push(
+        this.props.externalCodeChanges.subscribe(newCode => {
+          this.onCodeChanged(newCode);
+        })
+      );
     }
+    keymaster("f8, f9", () => {
+      this.onRunProgram();
+      return false;
+    });
+  }
 
-    render(): JSX.Element {
-        return (
-            <div className="code-panel-container">
-                {React.createElement(CodeInputLogoComponent, { ...this.props.codeInputProps, className: "code-input-container" })}
+  componentWillUnmount() {
+    this.subscriptions.forEach(s => s.unsubscribe());
+    keymaster.unbind("f8, f9");
+  }
+
+  onRunProgram = (): void => {
+    this.props.executionService.executeProgram(this.state.code);
+  };
+
+  render(): JSX.Element {
+    const execService = this.props.executionService;
+    return (
+      <div className="code-panel-container">
+        {this.renderSaveModal()}
+        {this.state.screenshotDataToSave && (
+          <ShareScreenshotModalComponent
+            imageBase64={this.state.screenshotDataToSave}
+            onClose={() => {
+              this.setState({ screenshotDataToSave: "" });
+            }}
+          />
+        )}
+
+        <ProgramControlsMenuComponent
+          isRunning={execService.isRunning}
+          existingProgramName={this.props.program.name}
+          runProgram={this.onRunProgram}
+          stopProgram={execService.stopProgram}
+          exportImage={this.exportScreenshot}
+          saveAsNew={this.showSaveDialog}
+          revertChanges={
+            this.state.hasLocalTempChanges && this.props.program.id ? this.revertCurrentProgram : undefined
+          }
+        />
+        <CodeInputLogoComponent
+          className="code-input-container"
+          editorTheme={this.props.editorTheme}
+          code={this.state.code}
+          focusCommands={execService.focusCommands}
+          onChanged={this.onCodeChanged}
+          onHotkey={this.onRunProgram}
+        />
+        {this.state.hasLocalTempChanges &&
+          this.props.program.id &&
+          !this.props.doNotShowLocalChangesIndicator && (
+            <div className="alert-not-stored-container">
+              <AlertMessageComponent message={_T("You have local changes in this program. ")} type="warning" />
             </div>
-        );
+          )}
+      </div>
+    );
+  }
+
+  onCodeChanged = (newCode: string) => {
+    this.setState({ code: newCode });
+    if (this.saveTempCodeTimer) {
+      clearTimeout(this.saveTempCodeTimer);
     }
+    this.saveTempCodeTimer = setTimeout(() => {
+      this.managementService.saveTempProgram(this.props.program.id, newCode);
+      if (!this.state.hasLocalTempChanges) {
+        this.setState({ hasLocalTempChanges: true });
+      }
+    }, 500);
+  };
+
+  renderSaveModal(): JSX.Element | null {
+    if (this.state.isSaveModalActive) {
+      return (
+        <SaveProgramModalComponent
+          programName={this.props.program.name}
+          onClose={() => {
+            this.setState({ isSaveModalActive: false });
+          }}
+          onSave={this.saveProgramAsCallback}
+        />
+      );
+    }
+    return null;
+  }
+
+  showSaveDialog = () => {
+    this.setState({ isSaveModalActive: true });
+  };
+
+  saveProgramAsCallback = async (newProgramName: string): Promise<void> => {
+    const screenshot = await this.props.executionService.getScreenshot(true);
+    const newProgram = await this.managementService.saveProgramToLibrary(
+      newProgramName,
+      screenshot,
+      this.state.code,
+      this.props.program
+    );
+    this.notificationService.push({
+      type: "info",
+      title: "Success",
+      message: _T("Program has been saved in the library.")
+    });
+    this.setState({ hasLocalTempChanges: false });
+    if (this.props.navigateAutomaticallyAfterSaveAs) {
+      this.navigationService.navigate({
+        route: Routes.playgroundCode.build({
+          programId: newProgram.id,
+          storageType: ProgramStorageType.gallery
+        })
+      });
+    }
+  };
+
+  revertCurrentProgram = async () => {
+    const code = await this.managementService.revertLocalTempChanges(this.props.program);
+    this.setState({ hasLocalTempChanges: false, code: code });
+  };
+
+  exportScreenshot = async () => {
+    const data = await this.props.executionService.getScreenshot(false);
+    if (!data) {
+      this.notificationService.push({
+        title: _T("Message"),
+        message: _T("Screenshot is not available because program has not been executed yet."),
+        type: "primary"
+      });
+    }
+    this.setState({ screenshotDataToSave: data });
+  };
 }
