@@ -1,78 +1,92 @@
-import * as React from 'react';
-import * as cn from 'classnames';
-
-import { Button, Nav, Navbar, NavDropdown, MenuItem, NavItem, DropdownButton } from 'react-bootstrap';
-import { LinkContainer } from 'react-router-bootstrap';
+import * as React from "react";
 import { _T } from "app/services/customizations/localization.service";
 
-interface IComponentState {
-}
+interface IComponentState {}
 
 interface IComponentProps {
-    className?: string
-    isRunning: boolean
-    runProgram: () => void
-    stopProgram: () => void
-    existingProgramName?: string
-    saveCurrent?: () => void
-    saveAsNew?: () => void
-    exportImage: () => void
+  className?: string;
+  isRunning: boolean;
+  runProgram: () => void;
+  stopProgram: () => void;
+  existingProgramName?: string;
+  revertChanges?: () => void;
+  saveAsNew?: () => void;
+  exportImage: () => void;
 }
 
 export class ProgramControlsMenuComponent extends React.Component<IComponentProps, IComponentState> {
-    constructor(props: IComponentProps) {
-        super(props);
+  constructor(props: IComponentProps) {
+    super(props);
 
-        this.state = {
-        }
-    }
+    this.state = {};
+  }
 
-    componentDidMount() {
-    }
-
-    render(): JSX.Element {
-        return <Nav className={cn(this.props.className)}>
-            <NavItem disabled={this.props.isRunning} onClick={this.props.runProgram}>
-                <span className="glyphicon glyphicon-play" aria-hidden="true"></span>
-                <span>&nbsp;</span>
-                <span>{_T("Run")}</span>
-            </NavItem>
-            <NavItem disabled={!this.props.isRunning} onClick={this.props.stopProgram}>
-                <span className="glyphicon glyphicon-stop" aria-hidden="true"></span>
-                <span>&nbsp;</span>
-                <span>{_T("Stop")}</span>
-            </NavItem>
-
-            <NavDropdown id="main-playground-menu-options-dropdown" bsClass="dropdown" noCaret pullRight
-                title={
-                    <span className="glyphicon glyphicon-option-vertical" aria-hidden="true"></span> as any
-                }>
-                {
-                    this.props.existingProgramName && this.props.saveCurrent &&
-                    <MenuItem onClick={this.props.saveCurrent}>
-                        <span className="glyphicon glyphicon-save" aria-hidden="true"></span>
-                        <span>&nbsp;&nbsp;</span>
-                        <span>{_T("Save program '%s'", { value: this.props.existingProgramName })}</span>
-                    </MenuItem>
-                }
-                {
-                    this.props.saveAsNew &&
-                    <MenuItem onClick={this.props.saveAsNew}>
-                        <span className="glyphicon glyphicon-file" aria-hidden="true"></span>
-                        {
-                            this.props.existingProgramName
-                                ? <span><span>&nbsp;&nbsp;</span>{_T("Save as new...")}</span>
-                                : <span><span>&nbsp;&nbsp;</span>{_T("Save to Gallery...")}</span>
-                        }
-                    </MenuItem>
-                }
-                <MenuItem divider />
-                <MenuItem onClick={this.props.exportImage}>
-                    <span className="glyphicon glyphicon-camera" aria-hidden="true"></span>
-                    <span>&nbsp;&nbsp;</span>
-                    <span>{_T("Take Screenshot")}</span>
-                </MenuItem>
-            </NavDropdown>
-        </Nav>
-    }
+  render(): JSX.Element | null {
+    return (
+      <div className="program-controls-menu-component">
+        {!this.props.isRunning && (
+          <button
+            type="button"
+            className="button is-success is-borderless"
+            onClick={this.props.runProgram}
+            title={_T("Execute the program (F9)")}
+          >
+            {_T("Run")}
+          </button>
+        )}{" "}
+        {this.props.isRunning && (
+          <button
+            type="button"
+            className="button is-warning is-borderless"
+            onClick={this.props.stopProgram}
+            title={_T("Stop execution of the program")}
+          >
+            {_T("Stop")}
+          </button>
+        )}{" "}
+        <div className="dropdown is-right is-hoverable is-borderless">
+          <div className="dropdown-trigger">
+            <button className="button is-light" aria-haspopup="true" aria-controls="dropdown-menu6">
+              <i className="fa fa-ellipsis-h" aria-hidden="true" />
+            </button>
+          </div>
+          <div className="dropdown-menu" id="dropdown-menu6" role="menu">
+            <div className="dropdown-content">
+              {this.props.revertChanges && (
+                <a
+                  className="dropdown-item"
+                  onClick={() => {
+                    this.props.revertChanges && this.props.revertChanges();
+                  }}
+                >
+                  <i className="fa fa-undo" aria-hidden="true" />
+                  &nbsp;&nbsp;
+                  {_T("Revert changes")}
+                </a>
+              )}
+              {this.props.revertChanges && <hr className="dropdown-divider" />}
+              {this.props.saveAsNew && (
+                <a className="dropdown-item" onClick={this.props.saveAsNew}>
+                  <i className="fa fa-file-code-o" aria-hidden="true" />
+                  &nbsp;&nbsp;
+                  {_T("Save to personal library")}
+                </a>
+              )}
+              <a className="dropdown-item">
+                <i className="fa fa-share-alt" aria-hidden="true" />
+                &nbsp;&nbsp;
+                {_T("Share")}
+              </a>
+              <hr className="dropdown-divider" />
+              <a className="dropdown-item" onClick={this.props.exportImage}>
+                <i className="fa fa-camera" aria-hidden="true" />
+                &nbsp;&nbsp;
+                {_T("Take screenshot")}
+              </a>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
 }
